move transaction errors to daemon crate. trim dependency features
