+++ cfdp-core/Cargo.toml
@@ -15,9 +15,8 @@
  pathdiff   = "~0.2"
  tempfile   = "~3.3"
  thiserror  = "~1.0"
- tokio      = { version = "1.28.2", features = ["macros", "net", "sync", "time"] }
+ tokio      = { version = "1.28.2", features = ["sync"] }
 
 [dev-dependencies]
  rstest       = "0.15.0"
  rstest_reuse = "0.5.0"
- tokio        = { version = "1.28.2", features = ["parking_lot", "macros"] }
