+++ cfdp-daemon/Cargo.toml
@@ -10,7 +10,7 @@
  cfdp-core   = { path = "../cfdp-core" }
  camino      = { version = "~1.1" }
  log         = '~0.4'
- tokio       = { version = "1.28.2", features = ["rt-multi-thread", "macros"] }
+ tokio       = { version = "1.28.2", features = ["time", "net", "rt-multi-thread", "macros"] }
  thiserror   = "~1.0"
